+++ src/engine/shapes/plane.rs
@@ -26,7 +26,7 @@ impl Plane {
     #[must_use]
     /// Retorna o ponto de interseção (de distância positiva) mais próximo entre um plano e um raio `r` \
     /// (`-INFINITY` se não há interseção)
-    pub fn intersects(&self, r: &Ray) -> (f32, Vec3) {
+    pub fn intersects(&self, r: &Ray) -> (f64, Vec3) {
         // Fórmula: n * (p - pc) = 0
         // n * (R(t) - pc) = 0
         // t = - n.dot(r.origin - pc) / n.dot(r.dr)
@@ -35,7 +35,7 @@ impl Plane {
         // se não, retorna o resultado da fórmula.
         let top = self.normal.dot(r.origin - self.pc);
         let bottom = self.normal.dot(r.dr);
-        if bottom == 0.0 { return (f32::NEG_INFINITY, Vec3::NULL) }
+        if bottom == 0.0 { return (f64::NEG_INFINITY, Vec3::NULL) }
         (-top/bottom, self.normal * -self.normal.dot(r.dr).signum())
     }
 }
